pyproject.toml
[project]
name = "kdense-session"
version = "0.1.0"
requires-python = "==3.12.10"
dependencies = [
    # Core scientific computing
    "numpy>=2.1.0",
    "pandas>=2.3.0",
    "matplotlib>=3.10.3",
    "scipy>=1.14.0",
    "seaborn>=0.13.0",
    "scikit-learn>=1.5.0",
    "statsmodels>=0.14.0",
    
    # Bioinformatics core
    "biopython>=1.85",
    
    # Single-cell analysis
    "anndata>=0.11.4",
    
    # Genomics and genetics
    "gget>=0.29.2",
    
    # Gene expression and enrichment
    "gseapy>=1.1.9",
    "pydeseq2>=0.5.2",
    
    # Data formats and utilities
    "pyyaml>=6.0.2",
    "beautifulsoup4>=4.13.4",
    "lxml>=6.0.0",
    "pillow>=11.0.0",
    "jinja2>=3.1.6",
    
    # Network requests (for API access to databases)
    "requests>=2.32.4",
    
    # Notebooks support
    "jupyter>=1.1.1",
    "ipykernel>=6.29.0",
    
    # Additional utilities
    "tqdm>=4.66.0",
    "plotly>=5.20.0",
    
    # Environment management
    "python-dotenv>=0.9.9",
]
